Renaming apiRequestError to apiError
