--- src/payloads/heartbeat-response.payload.ts
@@ -1,6 +1,0 @@
-export class HeartbeatResponsePayload {
-  lastBeatGeneratedAt: Date;
-  constructor(lastBeatGeneratedAt: Date) {
-    this.lastBeatGeneratedAt = lastBeatGeneratedAt;
-  }
-}
